% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DIdata.R
\docType{data}
\name{DIdata}
\alias{DIdata}
\title{Example and Test Data for drsmooth - dichotomous}
\format{Dose and response columns to be tested and modeled.
DIdata contains 3 numeric columns, each with a title:

Dose
    The dose tested.

n
    The number of cases receiving the dose.

Tumor
    The number of cases observed at the dose level.}
\usage{
DIdata
}
\description{
A .csv file containing a dose column "Dose", a column "n" defining how many cases received that dose,
and an outcome column "Tumor" indicating how many of cases of the outcome were observed at the dose level.
}
\keyword{internal}

