% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dosefactor.r
\name{dosefactor}
\alias{dosefactor}
\title{dosefactor}
\usage{
dosefactor(dosecolumn = "", data = NA)
}
\arguments{
\item{dosecolumn}{Name of dose column in dataframe.}

\item{data}{Input dataframe.}
}
\description{
Returns factorized dose column dose_fac in second column
}
\keyword{internal}

