% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drsmooth.print.r
\name{drsmooth.print}
\alias{drsmooth.print}
\title{Print Formatted Matrices for drsmooth}
\usage{
drsmooth.print(m)
}
\arguments{
\item{m}{test output matrices}
}
\description{
Print Formatted Matrices for drsmooth
}
\keyword{internal}

