\name{dunnetts.format}
\alias{dunnetts.format}
\title{Format Dunnett's Test Output}
\usage{
  dunnetts.format(summary, test, levels)
}
\arguments{
  \item{summary}{Summary object from dunnett's test for
  formatting}

  \item{test}{Name of the test to be formatted}

  \item{levels}{Dose factor levels defining rows in
  formatted output}
}
\description{
  Format Dunnett's Test Output
}
\keyword{internal}

