% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dunnetts.R
\name{dunnetts}
\alias{dunnetts}
\title{Dunnett's}
\usage{
dunnetts(targetcolumn, alternative, alpha, control, tot.obs, label, data)
}
\arguments{
\item{targetcolumn}{Character string, name of response column to be tested.}

\item{alternative}{Character string(s) specifying the direction of the alternative hypothesis.}

\item{alpha}{Significance level (numeric) to be used.}

\item{control}{Not relevant for this function}

\item{tot.obs}{Not relevant for this function}

\item{label}{Label of the alternative direction for output.}

\item{data}{Input dataframe.}
}
\description{
Dunnett's
}
\keyword{internal}

