\name{dunnetts1}
\alias{dunnetts1}
\title{Dunnett's One-tailed}
\usage{
  dunnetts1(targetcolumn, alpha, direction, data)
}
\arguments{
  \item{targetcolumn}{Character string, name of response
  column to be tested}

  \item{alpha}{Significance level (numeric) to be used}

  \item{direction}{Direction of the anticipated difference}

  \item{data}{Input dataframe.}
}
\description{
  Dunnett's One-tailed
}
\keyword{internal}

