\name{dunnetts2}
\alias{dunnetts2}
\title{Dunnett's Two-tailed}
\usage{
  dunnetts2(targetcolumn, alpha, direction, data)
}
\arguments{
  \item{targetcolumn}{Character string, name of response
  column to be tested}

  \item{alpha}{Significance level (numeric) to be used}

  \item{direction}{Direction of the anticipated difference}

  \item{data}{Input dataframe.}
}
\description{
  Dunnett's Two-tailed
}
\keyword{internal}

