\name{dunnettst3}
\alias{dunnettst3}
\title{Dunnett's T3 Test}
\usage{
  dunnettst3(targetcolumn, alpha, direction, data)
}
\arguments{
  \item{targetcolumn}{Character string, name of response
  column to be tested}

  \item{alpha}{Significance level (numeric) to be used}

  \item{direction}{Direction of the anticipated difference}

  \item{data}{Input dataframe.}
}
\description{
  Dunnett's T3 Test
}
\keyword{internal}

