% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dunns.format.r
\name{dunns.format}
\alias{dunns.format}
\title{Format Dunn's Test Output}
\usage{
dunns.format(model, test)
}
\arguments{
\item{model}{Dunn's test model object}

\item{test}{Name of the test executed}
}
\description{
Format Dunn's Test Output
}
\keyword{internal}

