% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dunns.R
\name{dunns}
\alias{dunns}
\title{Dunn's Test}
\usage{
dunns(targetcolumn, alternative, alpha, control, tot.obs, label, data)
}
\arguments{
\item{targetcolumn}{Character string, name of response column to be tested.}

\item{alternative}{The direction(s) of the alternative hypothesis.}

\item{alpha}{Significance level (numeric) to be used.}

\item{control}{Not used in this test.}

\item{tot.obs}{Not used in this test.}

\item{label}{Label indicating alternative.}

\item{data}{Input dataframe.}
}
\description{
Dunn's Test
}
\keyword{internal}

