\name{dunns1}
\alias{dunns1}
\title{Dunn's One-tailed}
\usage{
  dunns1(targetcolumn, alpha, direction, data)
}
\arguments{
  \item{targetcolumn}{Character string, name of response
  column to be tested}

  \item{alpha}{Significance level (numeric) to be used}

  \item{direction}{Direction of the anticipated difference}

  \item{data}{Input dataframe.}
}
\description{
  Dunn's One-tailed
}
\keyword{internal}

