\name{dunns2}
\alias{dunns2}
\title{Dunn's Two-tailed}
\usage{
  dunns2(targetcolumn, alpha, direction, data)
}
\arguments{
  \item{targetcolumn}{Character string, name of response
  column to be tested}

  \item{alpha}{Significance level (numeric) to be used}

  \item{direction}{Direction of the anticipated difference}

  \item{data}{Input dataframe.}
}
\description{
  Dunn's Two-tailed
}
\keyword{internal}

