% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firstDeriv.R
\name{firstDeriv}
\alias{firstDeriv}
\title{First Derivative Function(s)}
\usage{
firstDeriv(mod, n)
}
\arguments{
\item{mod}{The gam model.}

\item{n}{Prediction increments.}
}
\description{
First Derivative Function(s)
}
\keyword{internal}

