% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fishers.exact.R
\name{fishers.exact}
\alias{fishers.exact}
\title{Fisher's Exact Test}
\usage{
fishers.exact(targetcolumn, alternative, alpha, control, tot.obs, label, data)
}
\arguments{
\item{targetcolumn}{Character string, name of response column to be tested.}

\item{alternative}{Character string(s) specifying the direction of the alternative hypothesis.
Must be one or more of "greater", "two.sided", or "less".}

\item{alpha}{Significance level (numeric) to be used.}

\item{control}{Level of dose to be used as the control for dichotomous data.}

\item{tot.obs}{Total number of observations for each dose level for dichotomous data.}

\item{label}{Output label.}

\item{data}{Input dataframe.}
}
\description{
Fisher's Exact Test
}
\keyword{internal}

