% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg_prep.r
\name{pkg_prep}
\alias{new.prelimstats}
\alias{pkg_prep}
\alias{validate.prelimstats}
\title{PkgName Prep Functions}
\usage{
new.prelimstats()

validate.prelimstats(x)
}
\description{
PkgName Prep Functions
}
\keyword{internal}

