% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment.plot.r
\name{segment.plot}
\alias{segment.plot}
\title{Plot Segmented Test}
\usage{
segment.plot(mod = NA, dosecolumn = "", targetcolumn = "", data = NA)
}
\arguments{
\item{mod}{Segmented model object}

\item{dosecolumn}{Name of dose column.}

\item{targetcolumn}{Name of response column.}

\item{data}{Input dataframe.}
}
\description{
Plots the segmented function, if successful.
}
\keyword{internal}

