% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment.print.r
\name{segment.print}
\alias{segment.print}
\title{Print Segmented Statistics}
\usage{
segment.print(mod)
}
\arguments{
\item{mod}{Segmented model object}
}
\description{
Print Segmented Statistics
}
\keyword{internal}

