\docType{data}
\name{DRdata}
\alias{DRdata}
\title{Example and Test Data for drsmooth}
\format{Dose and response columns to be tested and modeled.
DRdata contains 4 numeric columns, each with a title:

dose
    The dose tested.

MF_Total
    The raw response column.

MF_Square
    The square root of the MF_Total column.

MF_Log
     The natural logarithm of the MF_Total column.}
\description{
  A .csv file containing one dose, one raw response column,
  and two transformed response columns.
}
\keyword{datasets}
\keyword{internal}

