% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Dose-response Modeling with Smoothing Splines}
\usage{
smooth(dosecolumn = "",
       targetcolumn = "",
       data = NA)
}
\arguments{
\item{dosecolumn}{Name of dose column of interest in dataframe.}

\item{targetcolumn}{Name of response column of interest in dataframe.}

\item{data}{Input dataframe.}
}
\description{
Replaced by drsmooth::drsmooth.
}
\details{
drsmooth::smooth has been replaced by drsmooth::drsmooth but a shell retained
in v.1.9.0 to prompt the switch.  v.2.0.0 will remove drsmooth::smooth altogether.
Please see help for drsmooth, where additional detail on the new functionality
is also available.
}

